Completely reformat MapOverlay It's a functional component now Removed map-deckgl-overlay Using BaseMap Using Portland data onClick works Co-authored-by: Jason Leonhard <devbrights@gmail.com> Co-authored-by: Doren Proctor <dorenproctor@gmail.com> Co-authored-by: David Mendoza <mendoza.line@aim.com>
