+++ packages/component-library/es/Logo/Logo.js
@@ -1,4 +1,5 @@
-import React, { PropTypes } from 'react';
+import React from 'react';
+import PropTypes from 'prop-types';
 import isClient from '../utils/isClient';
 
 var styles = {
