+++ packages/component-library/es/index.js
@@ -33,4 +33,8 @@ export { default as BaseMap } from './BaseMap/BaseMap';
 export { default as MapOverlay } from './MapOverlay/MapOverlay';
 export { default as HexOverlay } from './HexOverlay/HexOverlay';
 export { default as ScatterPlotMap } from './ScatterPlotMap/ScatterPlotMap';
+export { default as ScreenGridMap } from './ScreenGridMap/ScreenGridMap';
+export { default as PathMap } from './PathMap/PathMap';
+export { default as IconMap } from './IconMap/IconMap';
+
 import './fonts.css';
