+++ packages/component-library/src/MapOverlay/MapOverlay.js
@@ -2,93 +2,55 @@ import React, { Component } from 'react';
 import MapGL from 'react-map-gl';
 import { css } from 'emotion';
 import './mapbox-gl.css';
-import DeckGLOverlay from './map-deckgl-overlay.js';
 import PropTypes from 'prop-types';
+import DeckGL, {GeoJsonLayer} from 'deck.gl';
 
-const mapWrapper = css`
+
+const MapOverlay = (props) => {
+  const { viewport, data, mapboxStyle, mapboxToken, opacity, filled, wireframe, extruded, elevation, onLayerClick, getPosition } = props;
+  
+  const mapWrapper = css`
   margin: auto;
   max-width: 900px;
-`;
+  `;
 
-const colorScale = r => [r * 255, 140, 200 * (1 - r)];
+  const colorScale = r => [r * 255, 140, 200 * (1 - r)];
 
-const DATA_URL = 'https://raw.githubusercontent.com/uber-common/deck.gl-data/master/examples/geojson/vancouver-blocks.json'; // eslint-disable-line
+  const DATA_URL = 'https://raw.githubusercontent.com/uber-common/deck.gl-data/master/examples/geojson/vancouver-blocks.json'; // eslint-disable-line
 
-class MapOverlay extends Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      viewport:{
-        ...DeckGLOverlay.defaultViewport,
-        width: window.innerWidth > 900 ? 898 : window.innerWidth,
-        height: 400,
-        pitch: 45,
-        bearing: 0,
-        latitude: 49.254,   // vancouver bc
-        longitude: -123.13, // vancouver bc
-      },
-      data: null
+  const LIGHT_SETTINGS = {
+      lightsPosition: [-125, 50.5, 5000, -122.8, 48.5, 8000],
+      ambientRatio: 0.2,
+      diffuseRatio: 0.5,
+      specularRatio: 0.3,
+      lightsStrength: [1.0, 0.0, 2.0, 0.0],
+      numberOfLights: 2
     };
-
-    fetch(DATA_URL)
-      .then(resp => resp.json())
-      .then(data => this.setState({data}));
-    this.onViewportChange = this.onViewportChange.bind(this);
-    this.resize = this.resize.bind(this);
-  }
-
-  componentDidMount() {
-    window.addEventListener('resize', this.resize);
-    this.resize();
-  }
-
-  componentWillUnmount() {
-    window.removeEventListener('resize', this.resize);
-  }
-
-  onViewportChange(viewport) {
-    this.setState({
-      viewport: {...this.state.viewport, ...viewport},
-    });
-  }
-
-  resize() {
-    this.onViewportChange({
-      width: window.innerWidth > 900 ? 898 : window.innerWidth,
-      height: 450,
-    });
-  }
-
-  render() {
-    const {viewport, data} = this.state;
-    const { mapboxStyle, mapboxToken, opacity, filled, wireframe, extruded, elevation, onLayerClick, getPosition } = this.props;
-
-    return (
-      /*<div className={mapWrapper}>
-        <MapGL
-          className={'MapGL'}
-          {...viewport}
-             mapStyle={mapboxStyle}
-             onViewportChange={this.onViewportChange.bind(this)}
-             mapboxApiAccessToken={mapboxToken}
-             onViewportChange={ viewport => this.onViewportChange(viewport)}
-        >*/
-          <DeckGLOverlay
-            viewport={viewport}
-            data={data}
-            colorScale={colorScale}
-            opacity={opacity}
-            filled={filled}
-            wireframe={wireframe}
-            extruded={extruded}
-            elevation={elevation}
-            getPosition={getPosition}
-            onLayerClick={onLayerClick}
-          />
-       /* </MapGL>
-      </div>*/
-    );
-  };
+  
+  const layer = new GeoJsonLayer({
+    id: 'geojson',
+    data,
+    opacity: opacity,
+    stroked: false,
+    filled: filled,
+    extruded: extruded,
+    wireframe: wireframe,
+    fp64: true,
+    getElevation: f => Math.sqrt(f.properties.Shape_Length) * elevation,
+    getFillColor: f => colorScale(f.properties.Shape_Length),
+    getLineColor: f => [255, 255, 255],
+    lightSettings: LIGHT_SETTINGS,
+    pickable: true, // Boolean(this.props.onHover),
+    autoHighlight: true,
+    // onHover: this.props.onHover,
+    getPosition: getPosition,
+    onClick: onLayerClick,
+    // onHover: ({object}) => setTooltip(object.properties.name || object.properties.station)
+  });
+  
+  return (
+    <DeckGL {...viewport} layers={[layer]} />
+  );
 };
 
 MapOverlay.propTypes = {
@@ -104,4 +66,4 @@ MapOverlay.defaultProps = {
   mapboxStyle: "mapbox://styles/themendozaline/cjg6296ub04ot2sqv9izku3qq",
 };
 
-export default MapOverlay
+export default MapOverlay;
