+++ packages/component-library/stories/index.js
@@ -19,8 +19,8 @@ import dropdownStory from './DropdownMenu.story';
 import headerStory from './Header.story';
 import Welcome from './Welcome';
 import baseMapStory from './BaseMap.story';
-import mapOverlay from './MapOverlay.story';
-import hexOverlay from './HexOverlay.story';
+import mapOverlayStory from './MapOverlay.story';
+import hexOverlayStory from './HexOverlay.story';
 import scatterPlotMapStory from './ScatterPlotMap.story';
 import screenGridMapStory from './ScreenGridMap.story';
 import pathMapStory from './PathMap.story';
@@ -55,8 +55,8 @@ sankeyStory();
 sliderStory();
 dropdownStory();
 baseMapStory();
-mapOverlay();
-hexOverlay();
+mapOverlayStory();
+hexOverlayStory();
 scatterPlotMapStory();
 screenGridMapStory();
 pathMapStory();
